+++ packages/fullScreen/web/src/test/Fixtures/imagesContentState.ts
